fix: particles now working properly
