+++ package.json
@@ -15,8 +15,8 @@
     "@emotion/styled": "^11.11.0",
     "@fontsource/dm-sans": "^5.0.4",
     "@fontsource/montserrat": "^5.0.8",
+    "@tsparticles/engine": "^3.2.1",
     "@tsparticles/react": "^3.0.0",
-    "@tsparticles/react-workspace": "tsparticles/react",
     "@types/node": "^20.11.16",
     "@types/react": "^18.2.53",
     "@types/react-dom": "^18.2.18",
